pyproject.toml
[project]
name = "aicountant-be"
version = "0.1.0"
description = ""
authors = [
    {name = "sepehreghdami",email = "sep.eghdami@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "alembic (>=1.18.3,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "openai-agents (>=0.8.4,<0.9.0)"
]

[tool.poetry]
package-mode = false


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
